Conditionally define copy_proc_mem_*
Instead of marking `copy_proc_mem_syscall` and `copy_proc_mem_direct`
as `maybe_unused`, we can just define them only when `USE_DIRECT`
is either defined or not.

This also changes the order of defining the functions so that
they match whether or not `USE_DIRECT` is defined to avoid having
to do an `#ifndef`.